pyproject.toml
[project]
name = "qfeed"
version = "0.1.0"
description = "AI-powered technical interview practice platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.84.0",
    "boto3>=1.42.33",
    "colorama>=0.4.6",
    "elevenlabs>=2.36.0",
    "fastapi>=0.128.0",
    "google-genai>=1.60.0",
    "jupyterlab>=4.5.2",
    "kiwipiepy>=0.22.2",
    "korcen>=1.0.2",
    "langchain>=1.2.9",
    "langchain-core>=1.2.9",
    "langchain-google-genai>=4.2.0",
    "langfuse>=3.14.5",
    "langgraph>=1.0.8",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "pydantic-settings>=2.12.0",
    "pymongo>=4.16.0",
    "sentence-transformers>=5.2.2",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.0.0",
    "ruff>=0.14.13",
]

[tool.ruff]
ignore = ["E501"]  # 긴 줄 길이 오류 무시
line-length = 88

[tool.ruff.lint]
# "E": PEP 8 스타일 가이드를 위반한 형식 오류 검사
# "F": 코드 논리적 오류와 무의미한 코드 검사
# "S": 보안 취약점 검사
select = ["E", "F", "S"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
markers = [
    "unit: 유닛 테스트",
    "integration: 통합 테스트",
    "e2e: E2E 테스트",
    "slow: 느린 테스트",
]

[tool.ruff.lint.per-file-ignores]
# "test" 또는 "tests" 폴더 하위의 모든 파일에서는 S101 에러를 무시함
"tests/*" = ["S101"]
